Move Rust planning into core crate
