+++ .github/workflows/tests.yml
@@ -47,7 +47,7 @@ jobs:
           pnpm rust:test
       - name: Check fixture contracts
         run: |
-          cargo test -p agent-ci fixture_plan_contracts_match_snapshots
+          cargo test -p agent-ci-core fixture_plan_contracts_match_snapshots
           pnpm fixtures:check
       - name: Check native CLI golden output
         run: pnpm golden:cli
