+++ crates/agent-ci-core/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod expr;
 pub mod matrix;
+pub mod plan;
 pub mod workflow;
